+++ docs_src/src/guide/changelog.md
@@ -16,6 +16,7 @@ meta:
 
 - Added a way to customize the request's body field used by `JWTController` for the authentication process. (By default, "username" and "password" are used)
 - Added a new validation rule: `unique`.
+- Added `helper.Only()`.
 
 ## v3.1.0